NavBar functionality, and Beginning of Create Group
The NavBar links (Home, Tasks and Balances, and Messages) all redirect to their own page.

In Messages Page, the "Create Group" Button has been created and directs to its own page. In the Create Group Page, the it contains a friends list and the user can name the group.

Create Group Page:
- Cancel button -> Leads to Message page but alerts user that they have canceled creating a group.
- Create button -> leads to Message page but alerts user that they have created the group, with the name of their group.