Leaving Group and The Ability to see group Roster #5
On the messages.php, a user can go to a group chat and click "Settings" which will make a form appear on top of the group chat. This form shows the roster of the group chat and at the bottom shows an option to leave group. If the group is left then it is announced in the chat.